app/requirements.txt
black==22.10.0
Django==4.1.2
dj-database-url==0.5.0
djangorestframework==3.14.0
flake8==5.0.4
gunicorn==20.1.0
isort==5.10.1
psycopg2-binary==2.9.4
pytest-cov==4.0.0
pytest-django==4.5.2
pytest==7.2.0
whitenoise==6.2.0
